requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.3.0.post0
Pillow==10.1.0
PySocks==1.7.1
PyVirtualDisplay==3.0
requests==2.31.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.0.7
Werkzeug==3.0.1
wsproto==1.2.0
